refactor: clarify auto resource routing structure
