+++ reme/steps/file_io/_path.py
@@ -31,11 +31,6 @@ IMAGE_MIME_BY_EXT: dict[str, str] = {
 IMAGE_SUFFIXES = frozenset(IMAGE_MIME_BY_EXT)
 
 
-def is_image_file(path: str | Path) -> bool:
-    """Return True when ``path`` carries a known image file suffix."""
-    return Path(path).suffix.lower() in IMAGE_SUFFIXES
-
-
 _INVALID_CHARS = re.compile(r'[<>:"/\\|?*\x00-\x1f]')
 _RESERVED_NAMES = {
     "CON",
