+++ tests/integration/test_auto_resource.py
@@ -25,7 +25,7 @@ sys.path.insert(0, str(INTEGRATION_DIR))
 # pylint: disable=wrong-import-position
 from _workspace_fixture import workspace_env  # noqa: E402
 
-from reme.steps.evolve._auto_resource import _compute_note_stem  # noqa: E402
+from reme.steps.evolve.base_auto_resource import _compute_note_stem  # noqa: E402
 from reme.steps.evolve.auto_text_resource import _compute_agent_session_id  # noqa: E402
 
 RESOURCE_FILENAME = "project-roadmap.md"
